Add files via upload
This RWR bezel is unfinished for two main reasons. First that I got mine RWR from a simulator, which has obvious difference from the real one. I tried to fix it but didn't have exact dimensions. The second reason is that I don't have a cap on the brightness adjusting knob, therefore I can't model it. Please feel free to use it and modify it.